feat: Fix response structure
feat: Add series and seriesMovie modules
feat: Use transactions on batch operation

+++ source/routes/series/deleteSeries.controller.ts
@@ -8,39 +8,43 @@ export default function (request: FastifyRequest<{
 		seriesId: Series['id'];
 	};
 }>, reply: FastifyReply): void {
-	prisma['series'].findUnique({
-		select: {
-			userId: true
-		},
-		where: {
-			id: request['params']['seriesId']
-		}
-	})
-	.then(function (series: Pick<Series, 'userId'> | null): Promise<Prisma.BatchPayload> {
-		if(series !== null) {
-			if(series['userId'] === request['user']['id']) {
-				return prisma['series'].deleteMany({
-					where: {
-						id: request['params']['seriesId']
-					}
-				});
-			} else {
-				throw new Unauthorized('User must be same');
+	if(request['user']['isVerified']) {
+		prisma['series'].findUnique({
+			select: {
+				userId: true
+			},
+			where: {
+				id: request['params']['seriesId']
 			}
-		} else {
-			throw new NotFound('Parameter[\'seriesId\'] must be valid');
-		}
-	})
-	.then(function (result: Prisma.BatchPayload): void {
-		if(result['count'] === 1) {
-			reply.status(204).send();
-
-			return;
-		} else {
-			throw new NotFound('Parameter[\'seriesId\'] must be valid');
-		}
-	})
-	.catch(reply.send.bind(reply))
+		})
+		.then(function (series: Pick<Series, 'userId'> | null): Promise<Prisma.BatchPayload> {
+			if(series !== null) {
+				if(series['userId'] === request['user']['id']) {
+					return prisma['series'].deleteMany({
+						where: {
+							id: request['params']['seriesId']
+						}
+					});
+				} else {
+					throw new Unauthorized('User must be same');
+				}
+			} else {
+				throw new NotFound('Parameter[\'seriesId\'] must be valid');
+			}
+		})
+		.then(function (result: Prisma.BatchPayload): void {
+			if(result['count'] === 1) {
+				reply.status(204).send();
+	
+				return;
+			} else {
+				throw new NotFound('Parameter[\'seriesId\'] must be valid');
+			}
+		})
+		.catch(reply.send.bind(reply));
+	} else {
+		reply.send(new Unauthorized('User must be verified'));
+	}
 	
 	return;
 }
